+++ docs/conf.py
@@ -191,102 +191,29 @@ NumpyDocstring._parse = patched_parse
 
 # plot options
 
-# plot_include_source
+plot_include_source = False
+plot_html_show_source_link = False
+plot_html_show_formats = False
+plot_formats = ['png']
 # plot_pre_code
 # plot_basedir
-# plot_formats
 # plot_rcparams
 # plot_apply_rcparams
 # plot_working_directory
 
-# {% has_class = false -%}
-# {% for option in options -%}
-# {% if option.startswith(":class:") %}
-# {% has_class = true %}
-# {% endif %}
-# {% endfor %}
-
-
 plot_template = """
-{{ source_code }}
-
 {{ only_html }}
-
-   {% if source_link or (html_show_formats and not multi_image) %}
-   (
-   {%- if source_link -%}
-   `Source code <{{ source_link }}>`__
-   {%- endif -%}
-   {%- if html_show_formats and not multi_image -%}
-     {%- for img in images -%}
-       {%- for fmt in img.formats -%}
-         {%- if source_link or not loop.first -%}, {% endif -%}
-         `{{ fmt }} <{{ dest_dir }}/{{ img.basename }}.{{ fmt }}>`__
-       {%- endfor -%}
-     {%- endfor -%}
-   {%- endif -%}
-   )
-   {% endif %}
 
    {% for img in images %}
    {% set has_class = false %}
 
    .. figure:: {{ build_dir }}/{{ img.basename }}.{{ default_fmt }}
-      {% for option in options -%}
-      {%- if option.startswith(":class:") -%}
-      {%- set has_class = true -%}
-      {%- if "img-fluid" not in option -%}
-      {%- set option = option + " img-fluid" -%}
-      {%- endif -%}
-      {%- if "figure-img" not in option -%}
-      {%- set option = option + " figure-img" -%}
-      {%- endif -%}
-      {%- endif -%}
-      {{ option }}
-      {% endfor %}
-      {%- if not has_class -%}
       :class: figure-img img-fluid
-      {%- endif %}
-
-      {% if html_show_formats and multi_image -%}
-        (
-        {%- for fmt in img.formats -%}
-        {%- if not loop.first -%}, {% endif -%}
-        `{{ fmt }} <{{ dest_dir }}/{{ img.basename }}.{{ fmt }}>`__
-        {%- endfor -%}
-        )
-      {%- endif -%}
 
       {{ caption }}
-   {% endfor %}
-
-{{ only_latex }}
-
-   {% for img in images %}
-   {% if "pdf" in img.formats -%}
-   .. figure:: {{ build_dir }}/{{ img.basename }}.pdf
-      {% for option in options -%}
-      {{ option }}
-      {% endfor %}
-
-      {{ caption }}
-   {% endif -%}
-   {% endfor %}
-
-{{ only_texinfo }}
-
-   {% for img in images %}
-   .. image:: {{ build_dir }}/{{ img.basename }}.png
-      {% for option in options -%}
-      {{ option }}
-      {% endfor %}
 
    {% endfor %}
-
 """
-
-plot_html_show_source_link = False
-plot_html_show_formats = False
 
 # intersphinx options